resolved some problems after merging
