+++ src/user/user.controller.ts
@@ -5,9 +5,15 @@ import { AuthGuard } from '@nestjs/passport';
 @Controller('user')
 export class UserController {
   constructor(@Inject(UserService) private userService: UserService) {}
-  
+
   @Get('/:email')
   async findUserByEmail(@Param() email: string) {
     return this.userService.findUserByEmail(email);
+  }
+
+  @Get('/list/basic')
+  @UseGuards(AuthGuard('admin'))
+  async getStudentsBasicData() {
+    return this.userService.getStudentsBasicData();
   }
 }
